Uncommented analysis python file and notebook
Notebook is for Google collab. File paths must be changed. Comments are missing on most code and analysis and findings are summarized. Generally, findings were that it was impossible to reach a win percentage that was profitable in the long run. Given that there was nothing else to find, analysis was dropped midway.